pyproject.toml
[project]
name = "ir-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=1.4.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mwparserfromhell>=0.7.2",
    "networkx>=3.4.2",
    "nltk>=3.9.3",
    "pandas>=2.3.3",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
]
